Backend/requirements.txt
# Core (no build issues)
fastapi>=0.115.0
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.3

# Data validation
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator>=2.0.0

# Real-time / IoT Stream (Module 3)
websockets>=12.0

# NO numpy/pandas/scikit-learn YET - add later with conda
# Replace with simple dicts for hackathon MVP

# LLM integration
google-generativeai>=0.8.3

# Telegram & Async Requests
httpx>=0.27.0
python-telegram-bot>=21.0